# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bilu Review API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "summary": "List Reviews",
        "description": "返回审查历史列表（包括未完成的审查）。",
        "operationId": "list_reviews_api_v1_reviews_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Upload Review",
        "description": "上传文件并发起异步审查任务。\n\n在后台任务中执行：解析 → OCR → 模型审查 → 证据校验。\n前端通过轮询 GET /api/v1/reviews/{task_id} 获取进度。",
        "operationId": "create_upload_review_api_v1_reviews_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_review_api_v1_reviews_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/demos/{demo_id}": {
      "post": {
        "summary": "Create Demo Review",
        "description": "按服务端案例目录指定的执行模式发起演示审查。",
        "operationId": "create_demo_review_api_v1_reviews_demos__demo_id__post",
        "parameters": [
          {
            "name": "demo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Demo Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{task_id}": {
      "get": {
        "summary": "Review Detail",
        "description": "获取单条审查任务的完整详情（用于前端轮询进度与结果）。",
        "operationId": "review_detail_api_v1_reviews__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{task_id}/complete": {
      "post": {
        "summary": "Archive Review",
        "description": "完成复核 — 将审查任务置为已归档状态（只读）。",
        "operationId": "archive_review_api_v1_reviews__task_id__complete_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{task_id}/follow-ups": {
      "get": {
        "summary": "Review Follow Ups",
        "description": "返回被标记为\"补问清单\"的规则项列表。",
        "operationId": "review_follow_ups_api_v1_reviews__task_id__follow_ups_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{task_id}/report-data": {
      "get": {
        "summary": "Review Report Data",
        "description": "返回审查报告所需的全部数据（打印/导出用）。",
        "operationId": "review_report_data_api_v1_reviews__task_id__report_data_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{task_id}/results/{rule_id}/decision": {
      "patch": {
        "summary": "Save Decision",
        "description": "人工分流决策 — 对某条规则结果标记确认/补问/忽略。",
        "operationId": "save_decision_api_v1_reviews__task_id__results__rule_id__decision_patch",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_create_upload_review_api_v1_reviews_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_upload_review_api_v1_reviews_post"
      },
      "CompleteReviewResponse": {
        "properties": {
          "reviewStatus": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "archivedAt": {
            "type": "string",
            "title": "Archivedat"
          }
        },
        "type": "object",
        "required": [
          "reviewStatus",
          "archivedAt"
        ],
        "title": "CompleteReviewResponse"
      },
      "CreateTaskResponse": {
        "properties": {
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "type": "object",
        "required": [
          "taskId",
          "status"
        ],
        "title": "CreateTaskResponse",
        "description": "创建审查任务的即时响应（异步处理，需轮询最终结果）。"
      },
      "DecisionRequest": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ManualStatus"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DecisionRequest"
      },
      "DecisionResponse": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ReviewResult"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "DecisionResponse"
      },
...
      "DocumentPage": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "paragraphs": {
            "items": {
              "$ref": "#/components/schemas/DocumentParagraph"
            },
            "type": "array",
            "title": "Paragraphs"
          }
        },
        "type": "object",
        "required": [
          "page",
          "paragraphs"
        ],
        "title": "DocumentPage",
        "description": "文档中的一页，包含段落列表。"
      },
      "DocumentParagraph": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sourceType": {
            "$ref": "#/components/schemas/SourceType"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "text",
          "sourceType"
        ],
        "title": "DocumentParagraph",
        "description": "文档中的单个段落。\n\nAttributes:\n    text:       段落原文。\n    sourceType: 来源类型，决定如何渲染（文本/表格/图像识别）。\n    confidence: 图像识别的置信度（0-1），仅 VISION 类型有效。"
      },
      "EvidenceLocation": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "paragraph": {
            "type": "integer",
            "title": "Paragraph"
          }
        },
        "type": "object",
        "required": [
          "page",
          "paragraph"
        ],
        "title": "EvidenceLocation",
        "description": "证据在笔录中的位置（页号 + 段号，均从 1 开始计数）。"
      },
      "FollowUpListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewResult"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "FollowUpListResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualDecision": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ManualStatus",
            "default": "pending"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "title": "ManualDecision",
        "description": "办案人员对单条规则结果的人工处置记录。"
      },
      "ManualStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "supplemented",
          "ignored"
        ],
        "title": "ManualStatus",
        "description": "人工分流 — 办案人员对模型结果的最终处置状态。"
      },
      "ParsedDocument": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "format": {
            "type": "string",
            "enum": [
              "DOCX",
              "PDF",
              "SAMPLE"
            ],
            "title": "Format"
          },
          "pageCount": {
            "type": "integer",
            "title": "Pagecount"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/DocumentPage"
            },
            "type": "array",
            "title": "Pages"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sizeLabel": {
            "type": "string",
            "title": "Sizelabel"
          }
        },
        "type": "object",
        "required": [
          "name",
          "format",
          "pageCount",
          "pages",
          "text",
          "sizeLabel"
        ],
        "title": "ParsedDocument",
        "description": "标准化后的文档 — 统一 PDF/DOCX/SAMPLE 的内部表示。"
      },
...
      "ReportData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParsedDocument"
              },
              {
                "type": "null"
              }
            ]
          },
          "victimProfile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VictimProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "$ref": "#/components/schemas/ReviewMode"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "reviewStatus": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ReviewResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document",
          "victimProfile",
          "mode",
          "createdAt",
          "archivedAt",
          "reviewStatus",
          "results"
        ],
        "title": "ReportData",
        "description": "审查报告的全部数据 — 用于打印/导出。"
      },
      "ReviewListResponse": {
        "properties": {
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewSummary"
            },
            "type": "array",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": [
          "reviews"
        ],
        "title": "ReviewListResponse"
      },
      "ReviewMode": {
        "type": "string",
        "enum": [
          "qwen",
          "mock",
          "local"
        ],
        "title": "ReviewMode",
        "description": "审查结果的执行来源。"
      },
      "ReviewResult": {
        "properties": {
          "ruleId": {
            "type": "string",
            "title": "Ruleid"
          },
          "ruleName": {
            "type": "string",
            "title": "Rulename"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "status": {
            "$ref": "#/components/schemas/RuleStatus"
          },
          "missingFacts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missingfacts"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          },
          "evidenceLocation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceLocation"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidenceLocations": {
            "items": {
              "$ref": "#/components/schemas/EvidenceLocation"
            },
            "type": "array",
            "title": "Evidencelocations"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "suggestedQuestion": {
            "type": "string",
            "title": "Suggestedquestion"
          },
          "advisories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Advisories"
          },
          "manualDecision": {
            "$ref": "#/components/schemas/ManualDecision"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "ruleId",
          "ruleName",
          "category",
          "group",
          "status",
          "missingFacts",
          "evidence",
          "evidenceLocation",
          "reason",
          "suggestedQuestion",
          "source"
        ],
        "title": "ReviewResult",
        "description": "经过确定性校验后的最终规则审查结果（含人工分流状态）。"
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "in_review",
          "archived"
        ],
        "title": "ReviewStatus",
        "description": "审查复核状态 — 控制 UI 是否允许修改操作。"
      },
      "ReviewSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "documentName": {
            "type": "string",
            "title": "Documentname"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "reviewStatus": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "pendingCount": {
            "type": "integer",
            "title": "Pendingcount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "documentName",
          "createdAt",
          "updatedAt",
          "reviewStatus",
          "pendingCount"
        ],
        "title": "ReviewSummary",
        "description": "审查历史列表中的单条摘要。"
      },
      "ReviewTask": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "mode": {
            "$ref": "#/components/schemas/ReviewMode"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "parsing"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParsedDocument"
              },
              {
                "type": "null"
              }
            ]
          },
          "victimProfile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VictimProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ReviewResult"
            },
            "type": "array",
            "title": "Results"
          },
          "timings": {
            "$ref": "#/components/schemas/ReviewTimings"
          },
          "reviewStatus": {
            "$ref": "#/components/schemas/ReviewStatus",
            "default": "in_review"
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ReviewTask",
        "description": "审查任务 — 系统的核心聚合根，包含一份笔录的全部审查状态。"
      },
      "ReviewTimings": {
        "properties": {
          "parseMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsems"
          },
          "modelReviewMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelreviewms"
          },
          "modelGroupsMs": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Modelgroupsms"
          },
          "totalMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalms"
          }
        },
        "type": "object",
        "title": "ReviewTimings",
        "description": "审查任务各阶段的耗时统计（毫秒）。"
      },
...
      "RuleStatus": {
        "type": "string",
        "enum": [
          "covered",
          "missing",
          "incomplete",
          "not_applicable"
        ],
        "title": "RuleStatus",
        "description": "单条规则的审查结论状态。"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "native_text",
          "table",
          "vision"
        ],
        "title": "SourceType",
        "description": "段落内容来源类型 — 决定该段文本的可信度与展示方式。"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "uploading",
          "parsing",
          "recognizing",
          "checking",
          "validating",
          "completed",
          "failed"
        ],
        "title": "TaskStatus",
        "description": "审查任务状态机 — 按执行顺序严格流转。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VictimProfile": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 150.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "ethnicity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ethnicity"
          },
          "idNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idnumber"
          },
          "employer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employer"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact"
          }
        },
        "type": "object",
        "title": "VictimProfile",
        "description": "从笔录中提取的被害人基础信息，全部字段可为空（提取失败时）。"
      }
    }
  }
}